rename NewFacade to Facade
